Declarative, extensible scheduler
Key point: the scheduler job waits for all child jobs to complete, *as
well as* midnight the next day.  This avoids awkwardness when a job is
held for 24h or more.
The declarative job-ordering style makes adding more jobs easy, and can
be extended with conditionals etc if that seems worthwhile later.