Better error handling in job chain
Jobs only run if preceeding jobs exited with 0; except for the
scheduler.  Added "set -e" so any failing command in a script causes a
nonzero exit.  Added the "qsum -m a" (mail on abort) option, so the job
owner gets an email if any job is cancelled or exits with an error.